requirements.txt
base-model-guiosoft==0.0.3
cache-guiosoft==0.0.6
openpyxl==3.0.3
orjson==2.6.5
python-dateutil==2.8.1
python-dotenv==0.13.0
requests==2.23.0
